Move iOS APIGroups out of a class
