+++ introspy/IOS_Utils/Signature.py
@@ -1,5 +1,4 @@
-
-from .APIGroups import APIGroups
+from . import APIGroups
 
 
 class Signature(object):
